Update EIP-6404: Move to Draft
Merged by EIP-Bot.
