+++ EIPS/eip-6465.md
@@ -4,11 +4,11 @@ title: SSZ Withdrawals Root
 description: Migration of withdrawals MPT commitment to SSZ
 author: Etan Kissling (@etan-status), Mikhail Kalinin (@mkalinin)
 discussions-to: https://ethereum-magicians.org/t/eip-6465-ssz-withdrawals-root/12883
-status: Stagnant
+status: Draft
 type: Standards Track
 category: Core
 created: 2023-02-08
-requires: 2718, 4895, 6493
+requires: 2718, 4895, 6404
 ---
 
 ## Abstract
@@ -41,7 +41,7 @@ Definitions from existing specifications that are used throughout this document 
 | Name | Value |
 | - | - |
 | [`MAX_WITHDRAWALS_PER_PAYLOAD`](https://github.com/ethereum/consensus-specs/blob/ef434e87165e9a4c82a99f54ffd4974ae113f732/specs/capella/beacon-chain.md#execution) | `uint64(2**4)` (= 16) |
-| [`TRANSACTION_TYPE_SSZ`](./eip-6493.md#eip-2718-transaction-types) | `0x04` |
+| [`SSZ_TX_TYPE`](./eip-6404.md#networking) | `0x04` |
 
 ### SSZ `Withdrawal` container
 
@@ -95,7 +95,7 @@ typed-withdrawal = withdrawal-type || withdrawal-data
 
 When exchanging SSZ withdrawals via the [Ethereum Wire Protocol](https://github.com/ethereum/devp2p/blob/6b259a7003b4bfb18365ba690f4b00ba8a26393b/caps/eth.md), the following withdrawal envelope is used:
 
-- `Withdrawal`: `TRANSACTION_TYPE_SSZ || snappyFramed(ssz(Withdrawal))`
+- `Withdrawal`: `SSZ_TX_TYPE || snappyFramed(ssz(Withdrawal))`
 
 Objects are encoded using [SSZ](https://github.com/ethereum/consensus-specs/blob/ef434e87165e9a4c82a99f54ffd4974ae113f732/ssz/simple-serialize.md) and compressed using the Snappy framing format, matching the encoding of consensus objects as defined in the [consensus networking specification](https://github.com/ethereum/consensus-specs/blob/ef434e87165e9a4c82a99f54ffd4974ae113f732/specs/phase0/p2p-interface.md#ssz-snappy-encoding-strategy). As part of the encoding, the uncompressed object length is emitted; the RECOMMENDED limit to enforce per object is `8 + 8 + 20 + 8` (= 44) bytes.
