+++ assets/eip-6404/convert_tests.py
@@ -1,93 +1,68 @@
 from dataclasses import dataclass
 from convert import *
+from tx_hashes import compute_tx_hash
 
 @dataclass
 class Test:
+    tx_profile: Type[Profile]
     rlp_tx_hash: Hash32
     rlp_tx_bytes: bytes
-    rlp_receipt_bytes: bytes
-    ssz_tx_root: Root
+    ssz_tx_root: Hash32
     ssz_tx_bytes: bytes
-    ssz_receipt_bytes: bytes
 
 tests = [
     Test(
+        tx_profile=RlpLegacyTransaction,
         rlp_tx_hash=Hash32(bytes.fromhex("5c504ed432cb51138bcf09aa5e8a410dd4a1e204ef84bfed1be16dfba1b22060")),
         rlp_tx_bytes=bytes.fromhex("f86780862d79883d2000825208945df9b87991262f6ba471f09758cde1c0fc1de734827a69801ca088ff6cf0fefd94db46111149ae4bfc179e9b94721fffd821d38d16464b3f71d0a045e0aff800961cfce805daef7016b9b675c137a6a41a548f7b60a3484c06a33a"),
-        rlp_receipt_bytes=bytes.fromhex("f90128a096a8e009d2b88b1483e6941e6812e32263b05683fac202abc622a3e31aed1957825208b9010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000c0"),
-        ssz_tx_root=Root(bytes.fromhex("d8aa0c09b136ea863cd5dd3fe21d72226aaa3f30ac6ab56de1ea8f0a61b100e2")),
+        ssz_tx_root=Hash32(bytes.fromhex("d8aa0c09b136ea863cd5dd3fe21d72226aaa3f30ac6ab56de1ea8f0a61b100e2")),
         ssz_tx_bytes=bytes.fromhex("080000007b000000fd0000000000000000000000004d00000008520000000000005df9b87991262f6ba471f09758cde1c0fc1de734697a0000000000000000000000000000000000000000000000000000000000006f000000010000203d88792d00000000000000000000000000000000000000000000000000000300a1e4380a3b1f749673e270229993ee55f35663b488ff6cf0fefd94db46111149ae4bfc179e9b94721fffd821d38d16464b3f71d045e0aff800961cfce805daef7016b9b675c137a6a41a548f7b60a3484c06a33a01"),
-        ssz_receipt_bytes=bytes.fromhex("1b00000096a8e009d2b88b1483e6941e6812e32263b05683fac202abc622a3e31aed19570852000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000002c010000"),
     ),
     Test(
+        tx_profile=RlpLegacyTransaction,
         rlp_tx_hash=Hash32(bytes.fromhex("e9e91f1ee4b56c0df2e9f06c2b8c27c6076195a88a7b8537ba8313d80e6f124e")),
         rlp_tx_bytes=bytes.fromhex("f86e8243eb850df847580082c35094df190dc7190dfba737d7777a163445b7fff161338806113a84987be800801ca03b08715b4403c792b8c7567edea634088bedcd7f60d9352b1f16c69830f3afd5a010b9afb67d2ec8b956f0e1dbc07eb79152904f3a7bf789fc869db56320adfe09"),
-        rlp_receipt_bytes=bytes.fromhex("f901080182a410b9010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000c0"),
-        ssz_tx_root=Root(bytes.fromhex("78526295a3926d954a455a06a7eff6c84098b14294d4c8bfc3303d1836ace217")),
+        ssz_tx_root=Hash32(bytes.fromhex("78526295a3926d954a455a06a7eff6c84098b14294d4c8bfc3303d1836ace217")),
         ssz_tx_bytes=bytes.fromhex("080000007b000000fd00000000eb430000000000004d00000050c3000000000000df190dc7190dfba737d7777a163445b7fff1613300e87b98843a11060000000000000000000000000000000000000000000000006f0000000100005847f80d000000000000000000000000000000000000000000000000000000030032be343b94f860124dc4fee278fdcbd38c102d883b08715b4403c792b8c7567edea634088bedcd7f60d9352b1f16c69830f3afd510b9afb67d2ec8b956f0e1dbc07eb79152904f3a7bf789fc869db56320adfe0901"),
-        ssz_receipt_bytes=bytes.fromhex("3a0000000852000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000d01000001"),
     ),
     Test(
+        tx_profile=RlpLegacyTransaction,
         rlp_tx_hash=Hash32(bytes.fromhex("dc81918bf78322ce017c592e81c855f40bb96bd82da9779167de1de109962be6")),
         rlp_tx_bytes=bytes.fromhex("f88b822ecd8509839089a083015f909484654be796dad370032391d5479f8f1fd9ddd14e80a4d508e62389272d541e4168e5303b5838dcef5f9ac769a057f8dc6aabce7ec1e69b7ce5e625a078e88f9c69d217c76d92d5472f5812501021342f0054178c0579ee532dc2a218a0484723f0933633a6c213fc201ee3aecbe919579377b11a79acd887950a068e33"),
-        rlp_receipt_bytes=bytes.fromhex("f90109018301a93eb9010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000c0"),
-        ssz_tx_root=Root(bytes.fromhex("5004f2f523d0599fea4be4f933af406f7ade1d4973fe76dccf23b16c20789a55")),
+        ssz_tx_root=Hash32(bytes.fromhex("5004f2f523d0599fea4be4f933af406f7ade1d4973fe76dccf23b16c20789a55")),
         ssz_tx_bytes=bytes.fromhex("08000000a7000000ff000000000100000000000000cd2e00000000000055000000905f01000000000084654be796dad370032391d5479f8f1fd9ddd14e0000000000000000000000000000000000000000000000000000000000000000770000000100a089908309000000000000000000000000000000000000000000000000000000d508e62389272d541e4168e5303b5838dcef5f9ac769a057f8dc6aabce7ec1e69b7ce5e60300be49bd130e126a917ddb5fabf7cdeb6dd9887f4078e88f9c69d217c76d92d5472f5812501021342f0054178c0579ee532dc2a218484723f0933633a6c213fc201ee3aecbe919579377b11a79acd887950a068e3300"),
-        ssz_receipt_bytes=bytes.fromhex("3a0000002e05010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000d01000001"),
     ),
     Test(
+        tx_profile=RlpAccessListTransaction,
         rlp_tx_hash=Hash32(bytes.fromhex("8135b5403ea5528341d661cdadd8eb67983909fc1644b0a133de708b13e10937")),
         rlp_tx_bytes=bytes.fromhex("01f90cc30182153680830bb224943765521db364ee269e9540970fd21e5a3e82569980b905845f53273d00000000000000000000000000000000000000000000000000112cfa39a81d7c00000000000000000000000088e6a0c2ddd26feeb64f039a2c41296fcb3f5640000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000004e4128acb080000000000000000000000003765521db364ee269e9540970fd21e5a3e825699000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000001dac7be33000000000000000000000000000000000000000000000000000000010005bd8200000000000000000000000000000000000000000000000000000000000000a000000000000000000000000000000000000000000000000000000000000004200000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000010000000000000000000000006a9850e46518231b23e50467c975fa94026be5d5000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000324128acb0800000000000000000000000088e6a0c2ddd26feeb64f039a2c41296fcb3f564000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000001ade409ce94c21807e0000000000000000000000000000000000000000000000000000000010005bd8200000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000260000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000002000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2000000000000000000000000f1f85b2c54a2bd284b1cf4141d64fd171bd855390000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000c00000000000000000000000000000000000000000000000000000000000000044a9059cbb000000000000000000000000f1f85b2c54a2bd284b1cf4141d64fd171bd8553900000000000000000000000000000000000000000000000034a2e5498f6f66f80000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000a4022c0d9f0000000000000000000000000000000000000000000001ade409ce94c21807e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006a9850e46518231b23e50467c975fa94026be5d500000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000f906d7f859941c86b3cdf2a60ae3a574f7f71d44e2c50bddb87ef842a00000000000000000000000000000000000000000000000000000000000000003a050f2c0a7453ed72f1987f5c26e34dd531d584edf35a9b2dd71d8cbd0b46cbec2f87a94c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2f863a0c5192a5360898ecb25d40688e79d4568e38430536a3f38f85b3859328d3cedbca0cc805915145ea1d9dd6fe22fe624a946d6b02968a5cf3e32ab5376b3fcedd207a0390f6178407c9b8e95802b8659e6df8e34c1e3d4f8d6a49e6132bbcdd937b63af7943765521db364ee269e9540970fd21e5a3e825699e1a00000000000000000000000000000000000000000000000000000000000000000f8fe94a0b86991c6218b36c1d19d4a2e9eb0ce3606eb48f8e7a010d6a54a4754c8869d6886b5f5d7fbfa5b4522237ea5c60d11bc4e7a1ff9390ba07050c9e0f4ca769c69bd3a8ef740bc37934f8e2c036e5a723fd8ee048ed3f8c3a01f21a62c4538bacf2aabeca410f0fe63151869f172e03c0e00357ba26a341effa00000000000000000000000000000000000000000000000000000000000000001a0faa1053f6d4d1d5c438dc7bcb1d301a860f75f57f936878518deaed1f60940d0a0390f6178407c9b8e95802b8659e6df8e34c1e3d4f8d6a49e6132bbcdd937b63aa0441b08c38c1e5e3510dc0c6fdeb4e6a8d60744b906171767ae804339d85b866cf89b946a9850e46518231b23e50467c975fa94026be5d5f884a0918cfc45073e7a8befeda107cec8f3ce514f4a406e227c253b570e1d6485d524a00000000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000004a00000000000000000000000000000000000000000000000000000000000000001f79457ab1ec28d129707052df4df418d58a2d46d5f51e1a00000000000000000000000000000000000000000000000000000000000000002f87a9405a9cbe762b36632b3594da4f082340e0e5343e8f863a0faa1053f6d4d1d5c438dc7bcb1d301a860f75f57f936878518deaed1f60940d0a0c5192a5360898ecb25d40688e79d4568e38430536a3f38f85b3859328d3cedbca00000000000000000000000000000000000000000000000000000000000000002f85994613c773c7a1d85d2f1dcc051b0573d33470762ebf842a00563a8a19823933e751ef690567f0351d13ee18500841743a77290b3deeac37da041feecc8dc55520132a464384eb4cad04fca8408899b069cf80a59042ac84c7af85994c757acba3c0506218b3022266a9dc7f3612d85f5f842a0ad77bf801726b2c83e9e3fa10f319bf2df99397a561126e8c3df0f7596519f75a0f0afdce5671e49e66c5829c89ac80a06875a30b8a123df04dc6cc02b9c7c845cf8dd94f1f85b2c54a2bd284b1cf4141d64fd171bd85539f8c6a00000000000000000000000000000000000000000000000000000000000000008a00000000000000000000000000000000000000000000000000000000000000006a00000000000000000000000000000000000000000000000000000000000000007a00000000000000000000000000000000000000000000000000000000000000009a0000000000000000000000000000000000000000000000000000000000000000aa0000000000000000000000000000000000000000000000000000000000000000cd6940000000000000000000000000000000000000000c0f89b9488e6a0c2ddd26feeb64f039a2c41296fcb3f5640f884a00000000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000004a00000000000000000000000000000000000000000000000000000000000000001a0a6172f64e80168333d3bc97ef096a7cb251d24be47156dff7a40f6163a6ab42bd694a2327a938febf5fec13bacfb16ae10ecbc4cbdcfc0f8dd944d8dbd193d89b7b506be5dc9db75b91da00d6a1df8c6a00000000000000000000000000000000000000000000000000000000000000004a0b79e191a25a1c10448c48948add5400e67bdba9b68cb8e4ab5ccc4025afe7bf6a0000000000000000000000000000000000000000000000000000000000000000ba0d8f0dca3c14b74279ad90a75338db7cccdd5b2448369aeb32a94693a1c4cde0da00000000000000000000000000000000000000000000000000000000000000005a00000000000000000000000000000000000000000000000000000000000000002f794545973f28950f50fc6c7f52aab4ad214a27c0564e1a097176f14e7115c17104cb26360d789269f7e452d9b8ac6a3b8945c07081b33ac01a0dc6b057ac2591f402a243c3b81e429e750442d0a86f0fc112cf2eaaac26d4b8da03e5cace679d3104fae7f63ef0e64e7f4be5cfae64f9e03aebc0e19e5de0f29f8"),
-        rlp_receipt_bytes=bytes.fromhex("01f907360183064f13b901000020000001000000000020008000000000000000000000000000000004000000000000000000000000000800000800000a0000000800200000000000400000000000080000100008080000080010002000000000000000000000010000000000000000000000000000000000000000000000000000000000000000100008000000000000000000000000000000000000000000000100000800000040000000000000000000002010000000000000000000000000000000000020000000080000000000020000000000000000000000000008000000000010000000000000000000002800000004000000100000000000000000000420008000000a0000000000f9062bf89b94c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2f863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa000000000000000000000000088e6a0c2ddd26feeb64f039a2c41296fcb3f5640a00000000000000000000000003765521db364ee269e9540970fd21e5a3e825699a000000000000000000000000000000000000000000000000034b819a76c9101ddf89b94a0b86991c6218b36c1d19d4a2e9eb0ce3606eb48f863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa00000000000000000000000006a9850e46518231b23e50467c975fa94026be5d5a000000000000000000000000088e6a0c2ddd26feeb64f039a2c41296fcb3f5640a000000000000000000000000000000000000000000000000000000001dac7be33f89b94c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2f863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa00000000000000000000000003765521db364ee269e9540970fd21e5a3e825699a0000000000000000000000000f1f85b2c54a2bd284b1cf4141d64fd171bd85539a000000000000000000000000000000000000000000000000034a2e5498f6f66f8f89b9457ab1ec28d129707052df4df418d58a2d46d5f51f863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa0000000000000000000000000f1f85b2c54a2bd284b1cf4141d64fd171bd85539a00000000000000000000000006a9850e46518231b23e50467c975fa94026be5d5a00000000000000000000000000000000000000000000001ade409ce94c21807e0f87994f1f85b2c54a2bd284b1cf4141d64fd171bd85539e1a01c411e9a96e071241c2f21f7726b17ae89e3cab4c78be50e062b03a9fffbbad1b8400000000000000000000000000000000000000000000444ae3b74db3b9d4e6c6800000000000000000000000000000000000000000000008597e1b00a2ca3cb8bf8fc94f1f85b2c54a2bd284b1cf4141d64fd171bd85539f863a0d78ad95fa46c994b6551d0da85fc275fe613ce37657fb8d5e3d130840159d822a00000000000000000000000003765521db364ee269e9540970fd21e5a3e825699a00000000000000000000000006a9850e46518231b23e50467c975fa94026be5d5b880000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000034a2e5498f6f66f80000000000000000000000000000000000000000000001ade409ce94c21807e00000000000000000000000000000000000000000000000000000000000000000f9011c946a9850e46518231b23e50467c975fa94026be5d5f863a0c42079f94a6350d7e6235f29174924f928cc2ac818eb64fed8004e115fbcca67a00000000000000000000000003765521db364ee269e9540970fd21e5a3e825699a000000000000000000000000088e6a0c2ddd26feeb64f039a2c41296fcb3f5640b8a00000000000000000000000000000000000000000000001ade409ce94c21807e0fffffffffffffffffffffffffffffffffffffffffffffffffffffffe253841cd0000000000000000000000000000000000000000000010d1979907b2b7b44fb8000000000000000000000000000000000000000000000020160626e11f50affdfffffffffffffffffffffffffffffffffffffffffffffffffffffffffffbc8cdf9011c9488e6a0c2ddd26feeb64f039a2c41296fcb3f5640f863a0c42079f94a6350d7e6235f29174924f928cc2ac818eb64fed8004e115fbcca67a00000000000000000000000003765521db364ee269e9540970fd21e5a3e825699a00000000000000000000000003765521db364ee269e9540970fd21e5a3e825699b8a000000000000000000000000000000000000000000000000000000001dac7be33ffffffffffffffffffffffffffffffffffffffffffffffffcb47e658936efe23000000000000000000000000000000000000555373cbb17fb17a13909e0f3a3b00000000000000000000000000000000000000000000000066d5e9eea910e5ef0000000000000000000000000000000000000000000000000000000000030ca3"),
-        ssz_tx_root=Root(bytes.fromhex("5f31e697f90e2898f50d956a2c555098936956da4fc99e699b81836aa467c444")),
+        ssz_tx_root=Hash32(bytes.fromhex("5f31e697f90e2898f50d956a2c555098936956da4fc99e699b81836aa467c444")),
         ssz_tx_bytes=bytes.fromhex("08000000ef0c0000ff01000001010000000000000036150000000000005900000024b20b00000000003765521db364ee269e9540970fd21e5a3e82569900000000000000000000000000000000000000000000000000000000000000007b000000ff050000010000000000000000000000000000000000000000000000000000000000000000005f53273d00000000000000000000000000000000000000000000000000112cfa39a81d7c00000000000000000000000088e6a0c2ddd26feeb64f039a2c41296fcb3f5640000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000004e4128acb080000000000000000000000003765521db364ee269e9540970fd21e5a3e825699000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000001dac7be33000000000000000000000000000000000000000000000000000000010005bd8200000000000000000000000000000000000000000000000000000000000000a000000000000000000000000000000000000000000000000000000000000004200000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000010000000000000000000000006a9850e46518231b23e50467c975fa94026be5d5000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000324128acb0800000000000000000000000088e6a0c2ddd26feeb64f039a2c41296fcb3f564000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000001ade409ce94c21807e0000000000000000000000000000000000000000000000000000000010005bd8200000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000260000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000002000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2000000000000000000000000f1f85b2c54a2bd284b1cf4141d64fd171bd855390000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000c00000000000000000000000000000000000000000000000000000000000000044a9059cbb000000000000000000000000f1f85b2c54a2bd284b1cf4141d64fd171bd8553900000000000000000000000000000000000000000000000034a2e5498f6f66f80000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000a4022c0d9f0000000000000000000000000000000000000000000001ade409ce94c21807e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000006a9850e46518231b23e50467c975fa94026be5d5000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000003c000000940000000c010000440100003c020000d40200000c03000084030000dc030000340400000c05000024050000bc050000d4050000ac0600001c86b3cdf2a60ae3a574f7f71d44e2c50bddb87e18000000000000000000000000000000000000000000000000000000000000000000000350f2c0a7453ed72f1987f5c26e34dd531d584edf35a9b2dd71d8cbd0b46cbec2c02aaa39b223fe8d0a0e5c4f27ead9083c756cc218000000c5192a5360898ecb25d40688e79d4568e38430536a3f38f85b3859328d3cedbccc805915145ea1d9dd6fe22fe624a946d6b02968a5cf3e32ab5376b3fcedd207390f6178407c9b8e95802b8659e6df8e34c1e3d4f8d6a49e6132bbcdd937b63a3765521db364ee269e9540970fd21e5a3e825699180000000000000000000000000000000000000000000000000000000000000000000000a0b86991c6218b36c1d19d4a2e9eb0ce3606eb481800000010d6a54a4754c8869d6886b5f5d7fbfa5b4522237ea5c60d11bc4e7a1ff9390b7050c9e0f4ca769c69bd3a8ef740bc37934f8e2c036e5a723fd8ee048ed3f8c31f21a62c4538bacf2aabeca410f0fe63151869f172e03c0e00357ba26a341eff0000000000000000000000000000000000000000000000000000000000000001faa1053f6d4d1d5c438dc7bcb1d301a860f75f57f936878518deaed1f60940d0390f6178407c9b8e95802b8659e6df8e34c1e3d4f8d6a49e6132bbcdd937b63a441b08c38c1e5e3510dc0c6fdeb4e6a8d60744b906171767ae804339d85b866c6a9850e46518231b23e50467c975fa94026be5d518000000918cfc45073e7a8befeda107cec8f3ce514f4a406e227c253b570e1d6485d52400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000157ab1ec28d129707052df4df418d58a2d46d5f5118000000000000000000000000000000000000000000000000000000000000000000000205a9cbe762b36632b3594da4f082340e0e5343e818000000faa1053f6d4d1d5c438dc7bcb1d301a860f75f57f936878518deaed1f60940d0c5192a5360898ecb25d40688e79d4568e38430536a3f38f85b3859328d3cedbc0000000000000000000000000000000000000000000000000000000000000002613c773c7a1d85d2f1dcc051b0573d33470762eb180000000563a8a19823933e751ef690567f0351d13ee18500841743a77290b3deeac37d41feecc8dc55520132a464384eb4cad04fca8408899b069cf80a59042ac84c7ac757acba3c0506218b3022266a9dc7f3612d85f518000000ad77bf801726b2c83e9e3fa10f319bf2df99397a561126e8c3df0f7596519f75f0afdce5671e49e66c5829c89ac80a06875a30b8a123df04dc6cc02b9c7c845cf1f85b2c54a2bd284b1cf4141d64fd171bd85539180000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000600000000000000000000000000000000000000000000000000000000000000070000000000000000000000000000000000000000000000000000000000000009000000000000000000000000000000000000000000000000000000000000000a000000000000000000000000000000000000000000000000000000000000000c00000000000000000000000000000000000000001800000088e6a0c2ddd26feeb64f039a2c41296fcb3f564018000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000001a6172f64e80168333d3bc97ef096a7cb251d24be47156dff7a40f6163a6ab42ba2327a938febf5fec13bacfb16ae10ecbc4cbdcf180000004d8dbd193d89b7b506be5dc9db75b91da00d6a1d180000000000000000000000000000000000000000000000000000000000000000000004b79e191a25a1c10448c48948add5400e67bdba9b68cb8e4ab5ccc4025afe7bf6000000000000000000000000000000000000000000000000000000000000000bd8f0dca3c14b74279ad90a75338db7cccdd5b2448369aeb32a94693a1c4cde0d00000000000000000000000000000000000000000000000000000000000000050000000000000000000000000000000000000000000000000000000000000002545973f28950f50fc6c7f52aab4ad214a27c05641800000097176f14e7115c17104cb26360d789269f7e452d9b8ac6a3b8945c07081b33ac03001fd34033240c95aabf73e186a94b9576c6dab81bdc6b057ac2591f402a243c3b81e429e750442d0a86f0fc112cf2eaaac26d4b8d3e5cace679d3104fae7f63ef0e64e7f4be5cfae64f9e03aebc0e19e5de0f29f801"),
-        ssz_receipt_bytes=bytes.fromhex("3a000000d5a50400000000000020000001000000000020008000000000000000000000000000000004000000000000000000000000000800000800000a0000000800200000000000400000000000080000100008080000080010002000000000000000000000010000000000000000000000000000000000000000000000000000000000000000100008000000000000000000000000000000000000000000000100000800000040000000000000000000002010000000000000000000000000000000000020000000080000000000020000000000000000000000000008000000000010000000000000000000002800000004000000100000000000000000000420008000000a00000000000d0100000120000000bc00000058010000f4010000900200000c0300000804000024050000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc21c0000007c000000ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef00000000000000000000000088e6a0c2ddd26feeb64f039a2c41296fcb3f56400000000000000000000000003765521db364ee269e9540970fd21e5a3e82569900000000000000000000000000000000000000000000000034b819a76c9101dda0b86991c6218b36c1d19d4a2e9eb0ce3606eb481c0000007c000000ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef0000000000000000000000006a9850e46518231b23e50467c975fa94026be5d500000000000000000000000088e6a0c2ddd26feeb64f039a2c41296fcb3f564000000000000000000000000000000000000000000000000000000001dac7be33c02aaa39b223fe8d0a0e5c4f27ead9083c756cc21c0000007c000000ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef0000000000000000000000003765521db364ee269e9540970fd21e5a3e825699000000000000000000000000f1f85b2c54a2bd284b1cf4141d64fd171bd8553900000000000000000000000000000000000000000000000034a2e5498f6f66f857ab1ec28d129707052df4df418d58a2d46d5f511c0000007c000000ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef000000000000000000000000f1f85b2c54a2bd284b1cf4141d64fd171bd855390000000000000000000000006a9850e46518231b23e50467c975fa94026be5d50000000000000000000000000000000000000000000001ade409ce94c21807e0f1f85b2c54a2bd284b1cf4141d64fd171bd855391c0000003c0000001c411e9a96e071241c2f21f7726b17ae89e3cab4c78be50e062b03a9fffbbad10000000000000000000000000000000000000000000444ae3b74db3b9d4e6c6800000000000000000000000000000000000000000000008597e1b00a2ca3cb8bf1f85b2c54a2bd284b1cf4141d64fd171bd855391c0000007c000000d78ad95fa46c994b6551d0da85fc275fe613ce37657fb8d5e3d130840159d8220000000000000000000000003765521db364ee269e9540970fd21e5a3e8256990000000000000000000000006a9850e46518231b23e50467c975fa94026be5d5000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000034a2e5498f6f66f80000000000000000000000000000000000000000000001ade409ce94c21807e000000000000000000000000000000000000000000000000000000000000000006a9850e46518231b23e50467c975fa94026be5d51c0000007c000000c42079f94a6350d7e6235f29174924f928cc2ac818eb64fed8004e115fbcca670000000000000000000000003765521db364ee269e9540970fd21e5a3e82569900000000000000000000000088e6a0c2ddd26feeb64f039a2c41296fcb3f56400000000000000000000000000000000000000000000001ade409ce94c21807e0fffffffffffffffffffffffffffffffffffffffffffffffffffffffe253841cd0000000000000000000000000000000000000000000010d1979907b2b7b44fb8000000000000000000000000000000000000000000000020160626e11f50affdfffffffffffffffffffffffffffffffffffffffffffffffffffffffffffbc8cd88e6a0c2ddd26feeb64f039a2c41296fcb3f56401c0000007c000000c42079f94a6350d7e6235f29174924f928cc2ac818eb64fed8004e115fbcca670000000000000000000000003765521db364ee269e9540970fd21e5a3e8256990000000000000000000000003765521db364ee269e9540970fd21e5a3e82569900000000000000000000000000000000000000000000000000000001dac7be33ffffffffffffffffffffffffffffffffffffffffffffffffcb47e658936efe23000000000000000000000000000000000000555373cbb17fb17a13909e0f3a3b00000000000000000000000000000000000000000000000066d5e9eea910e5ef0000000000000000000000000000000000000000000000000000000000030ca3"),
     ),
     Test(
+        tx_profile=RlpFeeMarketTransaction,
         rlp_tx_hash=Hash32(bytes.fromhex("ba363483b992ef59342094ab98b8523ed6c055b5788e6726140d2badf27bf6d6")),
         rlp_tx_bytes=bytes.fromhex("02f9034c01830f9b358085064a2d5a958302690c946b75d8af000000e20b7a7ddf000ba900b4009a808403e168d79b2b2f1a397217ae1de4a6e543858b0191c9213aadbcc4ed256de6ecf902c0f85994c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2f842a012231cd4c753cb5530a43a74c45106c24765e6f81dc8927d4f4be7e53315d5a8a0f3bba710d3566d618899e2756f47b6b6ab971f4e7a9b49c8676839ce70a3b615f8dd94397217ae1de4a6e543858b0191c9213aadbcc4edf8c6a0000000000000000000000000000000000000000000000000000000000000000ca00000000000000000000000000000000000000000000000000000000000000008a00000000000000000000000000000000000000000000000000000000000000006a00000000000000000000000000000000000000000000000000000000000000007a00000000000000000000000000000000000000000000000000000000000000009a0000000000000000000000000000000000000000000000000000000000000000af90183944ecfc56672c7630b84dac9c1f7407579715de155f9016ba0000000000000000000000000000000000000000000000000000000000000000fa038ed69781a961bb9622064044b9913b1581bddd02ed6468e0e8c3899ae4ff6cea029cb8bd4e192d16f51155329ce8b0f5eb88a1d9e4d3b93ce07efbac9e1c4d175a0a34640eb1c607d836ceb4af07c59befcec4cdb4943351c3a245f8712063aa3b6a0000000000000000000000000000000000000000000000000000000000000000aa00000000000000000000000000000000000000000000000000000000000000008a00e0b1ffc883156bcc50271d13adbf2142a3da019b453b028078482a0967df23ba0dc59cdc7b4ec63ea402af3d39d169ec1338c1dd00754c63deca3620ef11ce970a04eb8e6d369a44c627e68958caefca5b050389afd4410af55103da22160f32e6fa0e9c8ce3daf53241f9a1be37866116a52a57048d3c7f27faf9b4e8c6e2e50cf81a0489f94fa9c8ead25fac83530419e4f3ae96ec772b8179124ec3bf465ad8ddb2e01a0bcd30031cd0ee1132ec0fd4debd5157269adf0e478e4c4733d846b09f8891dcda02afb6af719e9b548c7d543e7698c5a9c15f97d7f126277d7178ff33670d84e08"),
-        rlp_receipt_bytes=bytes.fromhex("02f9045b018307ff02b9010000200000000000000000000080000000000000000000004000000000400000000000008000000020001000000000040002000000084000000000000000000000000000000000000000000008000000200000000040000000000000000000000000000200000000000000000000000000000000000000000080000010000000000000800000000000000000000000000000000000000000080000004000000000000000000000000000000000000000000000000000100000000000000000400000000002000000000000000000000000000004000000001000001000000000000000200000000000000000000000000000000000000000000000000000000000f90350f89b94c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2f863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa00000000000000000000000006b75d8af000000e20b7a7ddf000ba900b4009a80a0000000000000000000000000397217ae1de4a6e543858b0191c9213aadbcc4eda000000000000000000000000000000000000000000000000003e168d700000000f89b944ecfc56672c7630b84dac9c1f7407579715de155f863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa0000000000000000000000000397217ae1de4a6e543858b0191c9213aadbcc4eda00000000000000000000000004ecfc56672c7630b84dac9c1f7407579715de155a00000000000000000000000000000000000000000000001df1858999999999999f89b944ecfc56672c7630b84dac9c1f7407579715de155f863a0ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3efa0000000000000000000000000397217ae1de4a6e543858b0191c9213aadbcc4eda00000000000000000000000006b75d8af000000e20b7a7ddf000ba900b4009a80a000000000000000000000000000000000000000000000238ece93666666666667f87994397217ae1de4a6e543858b0191c9213aadbcc4ede1a01c411e9a96e071241c2f21f7726b17ae89e3cab4c78be50e062b03a9fffbbad1b8400000000000000000000000000000000000000000000ad652dd7967ecf01d441000000000000000000000000000000000000000000000000122a5b6292417bf7af8fc94397217ae1de4a6e543858b0191c9213aadbcc4edf863a0d78ad95fa46c994b6551d0da85fc275fe613ce37657fb8d5e3d130840159d822a00000000000000000000000006b75d8af000000e20b7a7ddf000ba900b4009a80a00000000000000000000000006b75d8af000000e20b7a7ddf000ba900b4009a80b880000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000003e168d70000000000000000000000000000000000000000000000000000256de6ec0000000000000000000000000000000000000000000000000000000000000000000000000000"),
-        ssz_tx_root=Root(bytes.fromhex("ddf3fdb24235f2df98b9c3b5444176ab135bf6240a821f12f18a75e1f7a46f86")),
+        ssz_tx_root=Hash32(bytes.fromhex("ddf3fdb24235f2df98b9c3b5444176ab135bf6240a821f12f18a75e1f7a46f86")),
         ssz_tx_bytes=bytes.fromhex("080000007c030000ff030000020100000000000000359b0f00000000005d0000000c690200000000006b75d8af000000e20b7a7ddf000ba900b4009a80d768e103000000000000000000000000000000000000000000000000000000007f0000009a0000004e0300000100955a2d4a060000000000000000000000000000000000000000000000000000002b2f1a397217ae1de4a6e543858b0191c9213aadbcc4ed256de6ec0c000000640000003c010000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc21800000012231cd4c753cb5530a43a74c45106c24765e6f81dc8927d4f4be7e53315d5a8f3bba710d3566d618899e2756f47b6b6ab971f4e7a9b49c8676839ce70a3b615397217ae1de4a6e543858b0191c9213aadbcc4ed18000000000000000000000000000000000000000000000000000000000000000000000c0000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000600000000000000000000000000000000000000000000000000000000000000070000000000000000000000000000000000000000000000000000000000000009000000000000000000000000000000000000000000000000000000000000000a4ecfc56672c7630b84dac9c1f7407579715de15518000000000000000000000000000000000000000000000000000000000000000000000f38ed69781a961bb9622064044b9913b1581bddd02ed6468e0e8c3899ae4ff6ce29cb8bd4e192d16f51155329ce8b0f5eb88a1d9e4d3b93ce07efbac9e1c4d175a34640eb1c607d836ceb4af07c59befcec4cdb4943351c3a245f8712063aa3b6000000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000080e0b1ffc883156bcc50271d13adbf2142a3da019b453b028078482a0967df23bdc59cdc7b4ec63ea402af3d39d169ec1338c1dd00754c63deca3620ef11ce9704eb8e6d369a44c627e68958caefca5b050389afd4410af55103da22160f32e6fe9c8ce3daf53241f9a1be37866116a52a57048d3c7f27faf9b4e8c6e2e50cf81489f94fa9c8ead25fac83530419e4f3ae96ec772b8179124ec3bf465ad8ddb2e010000000000000000000000000000000000000000000000000000000000000000000300ae2fc483527b8ef99eb5d9b44875f005ba1fae13bcd30031cd0ee1132ec0fd4debd5157269adf0e478e4c4733d846b09f8891dcd2afb6af719e9b548c7d543e7698c5a9c15f97d7f126277d7178ff33670d84e0801"),
-        ssz_receipt_bytes=bytes.fromhex("3a000000efaf010000000000002000000000000000000000800000000000000000000040000000004000000000000080000000200010000000000400020000000840000000000000000000000000000000000000000000080000002000000000400000000000000000000000000002000000000000000000000000000000000000000000800000100000000000008000000000000000000000000000000000000000000800000040000000000000000000000000000000000000000000000000001000000000000000004000000000020000000000000000000000000000040000000010000010000000000000002000000000000000000000000000000000000000000000000000000000000d0100000114000000b00000004c010000e801000064020000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc21c0000007c000000ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef0000000000000000000000006b75d8af000000e20b7a7ddf000ba900b4009a80000000000000000000000000397217ae1de4a6e543858b0191c9213aadbcc4ed00000000000000000000000000000000000000000000000003e168d7000000004ecfc56672c7630b84dac9c1f7407579715de1551c0000007c000000ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef000000000000000000000000397217ae1de4a6e543858b0191c9213aadbcc4ed0000000000000000000000004ecfc56672c7630b84dac9c1f7407579715de1550000000000000000000000000000000000000000000001df18589999999999994ecfc56672c7630b84dac9c1f7407579715de1551c0000007c000000ddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef000000000000000000000000397217ae1de4a6e543858b0191c9213aadbcc4ed0000000000000000000000006b75d8af000000e20b7a7ddf000ba900b4009a8000000000000000000000000000000000000000000000238ece93666666666667397217ae1de4a6e543858b0191c9213aadbcc4ed1c0000003c0000001c411e9a96e071241c2f21f7726b17ae89e3cab4c78be50e062b03a9fffbbad10000000000000000000000000000000000000000000ad652dd7967ecf01d441000000000000000000000000000000000000000000000000122a5b6292417bf7a397217ae1de4a6e543858b0191c9213aadbcc4ed1c0000007c000000d78ad95fa46c994b6551d0da85fc275fe613ce37657fb8d5e3d130840159d8220000000000000000000000006b75d8af000000e20b7a7ddf000ba900b4009a800000000000000000000000006b75d8af000000e20b7a7ddf000ba900b4009a80000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000003e168d70000000000000000000000000000000000000000000000000000256de6ec0000000000000000000000000000000000000000000000000000000000000000000000000000"),
     ),
     Test(
+        tx_profile=RlpFeeMarketTransaction,
         rlp_tx_hash=Hash32(bytes.fromhex("a3b805acacb25da412a44bd9612a73464292fc684a400ab54f0a9626f7d9c3f2")),
         rlp_tx_bytes=bytes.fromhex("02f911d401038419fa53ad850c87eaa4da830efb7a8080b9117a608060405260405162000eda38038062000eda83398101604081905262000026916200049d565b828162000036828260006200004d565b50620000449050826200008a565b505050620005d0565b6200005883620000e5565b600082511180620000665750805b1562000085576200008383836200012760201b620001791760201c565b505b505050565b7f7e644d79422f17c01e4894b5f4f588d331ebfa28653d42ae832dc59e38c9798f620000b562000156565b604080516001600160a01b03928316815291841660208301520160405180910390a1620000e2816200018f565b50565b620000f08162000244565b6040516001600160a01b038216907fbc7cd75a20ee27fd9adebab32041f755214dbc6bffa90cc0225b39da2e5c2d3b90600090a250565b60606200014f838360405180606001604052806027815260200162000eb360279139620002f8565b9392505050565b60006200018060008051602062000e9383398151915260001b6200037760201b620001a51760201c565b546001600160a01b0316919050565b6001600160a01b038116620001fa5760405162461bcd60e51b815260206004820152602660248201527f455243313936373a206e65772061646d696e20697320746865207a65726f206160448201526564647265737360d01b60648201526084015b60405180910390fd5b806200022360008051602062000e9383398151915260001b6200037760201b620001a51760201c565b80546001600160a01b0319166001600160a01b039290921691909117905550565b6200025a816200037a60201b620001a81760201c565b620002be5760405162461bcd60e51b815260206004820152602d60248201527f455243313936373a206e657720696d706c656d656e746174696f6e206973206e60448201526c1bdd08184818dbdb9d1c9858dd609a1b6064820152608401620001f1565b80620002237f360894a13ba1a3210667c828492db98dca3e2076cc3735a920a3ca505d382bbc60001b6200037760201b620001a51760201c565b6060600080856001600160a01b0316856040516200031791906200057d565b600060405180830381855af49150503d806000811462000354576040519150601f19603f3d011682016040523d82523d6000602084013e62000359565b606091505b5090925090506200036d8683838762000389565b9695505050505050565b90565b6001600160a01b03163b151590565b60608315620003fa578251620003f2576001600160a01b0385163b620003f25760405162461bcd60e51b815260206004820152601d60248201527f416464726573733a2063616c6c20746f206e6f6e2d636f6e74726163740000006044820152606401620001f1565b508162000406565b6200040683836200040e565b949350505050565b8151156200041f5781518083602001fd5b8060405162461bcd60e51b8152600401620001f191906200059b565b80516001600160a01b03811681146200045357600080fd5b919050565b634e487b7160e01b600052604160045260246000fd5b60005b838110156200048b57818101518382015260200162000471565b83811115620000835750506000910152565b600080600060608486031215620004b357600080fd5b620004be846200043b565b9250620004ce602085016200043b565b60408501519092506001600160401b0380821115620004ec57600080fd5b818601915086601f8301126200050157600080fd5b81518181111562000516576200051662000458565b604051601f8201601f19908116603f0116810190838211818310171562000541576200054162000458565b816040528281528960208487010111156200055b57600080fd5b6200056e8360208301602088016200046e565b80955050505050509250925092565b60008251620005918184602087016200046e565b9190910192915050565b6020815260008251806020840152620005bc8160408501602087016200046e565b601f01601f19169190910160400192915050565b6108b380620005e06000396000f3fe60806040523661001357610011610017565b005b6100115b61001f6101b7565b6001600160a01b0316336001600160a01b0316141561016f5760606001600160e01b031960003516631b2ce7f360e11b8114156100655761005e6101ea565b9150610167565b6001600160e01b0319811663278f794360e11b14156100865761005e610241565b6001600160e01b031981166308f2839760e41b14156100a75761005e610287565b6001600160e01b031981166303e1469160e61b14156100c85761005e6102b8565b6001600160e01b03198116635c60da1b60e01b14156100e95761005e6102f8565b60405162461bcd60e51b815260206004820152604260248201527f5472616e73706172656e745570677261646561626c6550726f78793a2061646d60448201527f696e2063616e6e6f742066616c6c6261636b20746f2070726f78792074617267606482015261195d60f21b608482015260a4015b60405180910390fd5b815160208301f35b61017761030c565b565b606061019e83836040518060600160405280602781526020016108576027913961031c565b9392505050565b90565b6001600160a01b03163b151590565b60007fb53127684a568b3173ae13b9f8a6016e243e63b6e8ee1178d6a717850b5d61035b546001600160a01b0316919050565b60606101f4610394565b600061020336600481846106a2565b81019061021091906106e8565b905061022d8160405180602001604052806000815250600061039f565b505060408051602081019091526000815290565b606060008061025336600481846106a2565b8101906102609190610719565b915091506102708282600161039f565b604051806020016040528060008152509250505090565b6060610291610394565b60006102a036600481846106a2565b8101906102ad91906106e8565b905061022d816103cb565b60606102c2610394565b60006102cc6101b7565b604080516001600160a01b03831660208201529192500160405160208183030381529060405291505090565b6060610302610394565b60006102cc610422565b610177610317610422565b610431565b6060600080856001600160a01b0316856040516103399190610807565b600060405180830381855af49150503d8060008114610374576040519150601f19603f3d011682016040523d82523d6000602084013e610379565b606091505b509150915061038a86838387610455565b9695505050505050565b341561017757600080fd5b6103a8836104d3565b6000825111806103b55750805b156103c6576103c48383610179565b505b505050565b7f7e644d79422f17c01e4894b5f4f588d331ebfa28653d42ae832dc59e38c9798f6103f46101b7565b604080516001600160a01b03928316815291841660208301520160405180910390a161041f81610513565b50565b600061042c6105bc565b905090565b3660008037600080366000845af43d6000803e808015610450573d6000f35b3d6000fd5b606083156104c15782516104ba576001600160a01b0385163b6104ba5760405162461bcd60e51b815260206004820152601d60248201527f416464726573733a2063616c6c20746f206e6f6e2d636f6e7472616374000000604482015260640161015e565b50816104cb565b6104cb83836105e4565b949350505050565b6104dc8161060e565b6040516001600160a01b038216907fbc7cd75a20ee27fd9adebab32041f755214dbc6bffa90cc0225b39da2e5c2d3b90600090a250565b6001600160a01b0381166105785760405162461bcd60e51b815260206004820152602660248201527f455243313936373a206e65772061646d696e20697320746865207a65726f206160448201526564647265737360d01b606482015260840161015e565b807fb53127684a568b3173ae13b9f8a6016e243e63b6e8ee1178d6a717850b5d61035b80546001600160a01b0319166001600160a01b039290921691909117905550565b60007f360894a13ba1a3210667c828492db98dca3e2076cc3735a920a3ca505d382bbc6101db565b8151156105f45781518083602001fd5b8060405162461bcd60e51b815260040161015e9190610823565b6001600160a01b0381163b61067b5760405162461bcd60e51b815260206004820152602d60248201527f455243313936373a206e657720696d706c656d656e746174696f6e206973206e60448201526c1bdd08184818dbdb9d1c9858dd609a1b606482015260840161015e565b807f360894a13ba1a3210667c828492db98dca3e2076cc3735a920a3ca505d382bbc61059b565b600080858511156106b257600080fd5b838611156106bf57600080fd5b5050820193919092039150565b80356001600160a01b03811681146106e357600080fd5b919050565b6000602082840312156106fa57600080fd5b61019e826106cc565b634e487b7160e01b600052604160045260246000fd5b6000806040838503121561072c57600080fd5b610735836106cc565b9150602083013567ffffffffffffffff8082111561075257600080fd5b818501915085601f83011261076657600080fd5b81358181111561077857610778610703565b604051601f8201601f19908116603f011681019083821181831017156107a0576107a0610703565b816040528281528860208487010111156107b957600080fd5b8260208601602083013760006020848301015280955050505050509250929050565b60005b838110156107f65781810151838201526020016107de565b838111156103c45750506000910152565b600082516108198184602087016107db565b9190910192915050565b60208152600082518060208401526108428160408501602087016107db565b601f01601f1916919091016040019291505056fe416464726573733a206c6f772d6c6576656c2064656c65676174652063616c6c206661696c6564a264697066735822122012bb4f564f73959a03513dc74fc3c6e40e8386e6f02c16b78d6db00ce0aa16af64736f6c63430008090033b53127684a568b3173ae13b9f8a6016e243e63b6e8ee1178d6a717850b5d6103416464726573733a206c6f772d6c6576656c2064656c65676174652063616c6c206661696c65640000000000000000000000007e22fcb742572515d1c3fef972ec066c995820ef000000000000000000000000d89f6892a2b58d50e0e8f51a961b12a25ce3f7570000000000000000000000000000000000000000000000000000000000000060000000000000000000000000000000000000000000000000000000000000020457405d050000000000000000000000005f4ec3df9cbd43714fe2740f5e3616155c5b8419000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec70000000000000000000000000000000000000000000000000000000064c7eb930000000000000000000000000000000000000000000000000000000064d2bde300000000000000000000000000000000000000000000000000000000000000e0000000000000000000000000000000000000000000000000000000003b9aca000000000000000000000000004580900be782a567ad407cc8fcde24dbda2a1aae000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000a000000000000000000000000000000000000000000000000000000000000000e000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000165a0bc00000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000009795118749000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000023a8f4b6a00000000000000000000000000000000000000000000000000000000c080a06fd00841d705c002a846de1b56d0c36941cd2897a7bb19419d427a252bc4fff4a05be5ed4725f8b6047b5309da5f47d1e98fa2c79e5c0841595291fae7167de108"),
-        rlp_receipt_bytes=bytes.fromhex("02f90354018316fa7cb901000000000000000000000000000000000040000000000000000080000001000000000000000000000000000000000000000000000200004000000000000201000000000000000000000000000000000200000100000000000000000000000000000000000002000000000000000000080004000080000000000000000000000040000000000000000000000004000000000000000000008000000000040080000000000000000000000000000000040000000000000000000000000000000000000000002020000000000000000004000a000000000400000000000000000020000000000000000000000000000000000000000000000000000000000000020000f90249f85a94451445a9adb7e4c8ab4a266755626c136e093a4ef842a0bc7cd75a20ee27fd9adebab32041f755214dbc6bffa90cc0225b39da2e5c2d3ba00000000000000000000000007e22fcb742572515d1c3fef972ec066c995820ef80f87b94451445a9adb7e4c8ab4a266755626c136e093a4ef863a08be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0a00000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000979db18107552faa36a52480a1dbb65ed3f51d7080f89994451445a9adb7e4c8ab4a266755626c136e093a4ee1a023f6ad8232d75562dd1c6b37dfc895af6bfc1ecd0fb3b88722c6a5e6b4dc9a20b8600000000000000000000000000000000000000000000000000000000064c7eb930000000000000000000000000000000000000000000000000000000064d2bde30000000000000000000000000000000000000000000000000000000064c7e277f85894451445a9adb7e4c8ab4a266755626c136e093a4ee1a07f26b83ff96e1f2b6a682f133852f6798a09c465da95921460cefb3847402498a00000000000000000000000000000000000000000000000000000000000000001f87994451445a9adb7e4c8ab4a266755626c136e093a4ee1a07e644d79422f17c01e4894b5f4f588d331ebfa28653d42ae832dc59e38c9798fb8400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000d89f6892a2b58d50e0e8f51a961b12a25ce3f757"),
-        ssz_tx_root=Root(bytes.fromhex("1b7474a5d934e0bd725f35a6bdcb5b5e33b18c4cce6cf20706438a6bc707c2c7")),
+        ssz_tx_root=Hash32(bytes.fromhex("1b7474a5d934e0bd725f35a6bdcb5b5e33b18c4cce6cf20706438a6bc707c2c7")),
         ssz_tx_bytes=bytes.fromhex("0800000013120000df0300000201000000000000000300000000000000490000007afb0e000000000000000000000000000000000000000000000000000000000000000000000000006b000000e5110000e51100000100daa4ea870c000000000000000000000000000000000000000000000000000000608060405260405162000eda38038062000eda83398101604081905262000026916200049d565b828162000036828260006200004d565b50620000449050826200008a565b505050620005d0565b6200005883620000e5565b600082511180620000665750805b1562000085576200008383836200012760201b620001791760201c565b505b505050565b7f7e644d79422f17c01e4894b5f4f588d331ebfa28653d42ae832dc59e38c9798f620000b562000156565b604080516001600160a01b03928316815291841660208301520160405180910390a1620000e2816200018f565b50565b620000f08162000244565b6040516001600160a01b038216907fbc7cd75a20ee27fd9adebab32041f755214dbc6bffa90cc0225b39da2e5c2d3b90600090a250565b60606200014f838360405180606001604052806027815260200162000eb360279139620002f8565b9392505050565b60006200018060008051602062000e9383398151915260001b6200037760201b620001a51760201c565b546001600160a01b0316919050565b6001600160a01b038116620001fa5760405162461bcd60e51b815260206004820152602660248201527f455243313936373a206e65772061646d696e20697320746865207a65726f206160448201526564647265737360d01b60648201526084015b60405180910390fd5b806200022360008051602062000e9383398151915260001b6200037760201b620001a51760201c565b80546001600160a01b0319166001600160a01b039290921691909117905550565b6200025a816200037a60201b620001a81760201c565b620002be5760405162461bcd60e51b815260206004820152602d60248201527f455243313936373a206e657720696d706c656d656e746174696f6e206973206e60448201526c1bdd08184818dbdb9d1c9858dd609a1b6064820152608401620001f1565b80620002237f360894a13ba1a3210667c828492db98dca3e2076cc3735a920a3ca505d382bbc60001b6200037760201b620001a51760201c565b6060600080856001600160a01b0316856040516200031791906200057d565b600060405180830381855af49150503d806000811462000354576040519150601f19603f3d011682016040523d82523d6000602084013e62000359565b606091505b5090925090506200036d8683838762000389565b9695505050505050565b90565b6001600160a01b03163b151590565b60608315620003fa578251620003f2576001600160a01b0385163b620003f25760405162461bcd60e51b815260206004820152601d60248201527f416464726573733a2063616c6c20746f206e6f6e2d636f6e74726163740000006044820152606401620001f1565b508162000406565b6200040683836200040e565b949350505050565b8151156200041f5781518083602001fd5b8060405162461bcd60e51b8152600401620001f191906200059b565b80516001600160a01b03811681146200045357600080fd5b919050565b634e487b7160e01b600052604160045260246000fd5b60005b838110156200048b57818101518382015260200162000471565b83811115620000835750506000910152565b600080600060608486031215620004b357600080fd5b620004be846200043b565b9250620004ce602085016200043b565b60408501519092506001600160401b0380821115620004ec57600080fd5b818601915086601f8301126200050157600080fd5b81518181111562000516576200051662000458565b604051601f8201601f19908116603f0116810190838211818310171562000541576200054162000458565b816040528281528960208487010111156200055b57600080fd5b6200056e8360208301602088016200046e565b80955050505050509250925092565b60008251620005918184602087016200046e565b9190910192915050565b6020815260008251806020840152620005bc8160408501602087016200046e565b601f01601f19169190910160400192915050565b6108b380620005e06000396000f3fe60806040523661001357610011610017565b005b6100115b61001f6101b7565b6001600160a01b0316336001600160a01b0316141561016f5760606001600160e01b031960003516631b2ce7f360e11b8114156100655761005e6101ea565b9150610167565b6001600160e01b0319811663278f794360e11b14156100865761005e610241565b6001600160e01b031981166308f2839760e41b14156100a75761005e610287565b6001600160e01b031981166303e1469160e61b14156100c85761005e6102b8565b6001600160e01b03198116635c60da1b60e01b14156100e95761005e6102f8565b60405162461bcd60e51b815260206004820152604260248201527f5472616e73706172656e745570677261646561626c6550726f78793a2061646d60448201527f696e2063616e6e6f742066616c6c6261636b20746f2070726f78792074617267606482015261195d60f21b608482015260a4015b60405180910390fd5b815160208301f35b61017761030c565b565b606061019e83836040518060600160405280602781526020016108576027913961031c565b9392505050565b90565b6001600160a01b03163b151590565b60007fb53127684a568b3173ae13b9f8a6016e243e63b6e8ee1178d6a717850b5d61035b546001600160a01b0316919050565b60606101f4610394565b600061020336600481846106a2565b81019061021091906106e8565b905061022d8160405180602001604052806000815250600061039f565b505060408051602081019091526000815290565b606060008061025336600481846106a2565b8101906102609190610719565b915091506102708282600161039f565b604051806020016040528060008152509250505090565b6060610291610394565b60006102a036600481846106a2565b8101906102ad91906106e8565b905061022d816103cb565b60606102c2610394565b60006102cc6101b7565b604080516001600160a01b03831660208201529192500160405160208183030381529060405291505090565b6060610302610394565b60006102cc610422565b610177610317610422565b610431565b6060600080856001600160a01b0316856040516103399190610807565b600060405180830381855af49150503d8060008114610374576040519150601f19603f3d011682016040523d82523d6000602084013e610379565b606091505b509150915061038a86838387610455565b9695505050505050565b341561017757600080fd5b6103a8836104d3565b6000825111806103b55750805b156103c6576103c48383610179565b505b505050565b7f7e644d79422f17c01e4894b5f4f588d331ebfa28653d42ae832dc59e38c9798f6103f46101b7565b604080516001600160a01b03928316815291841660208301520160405180910390a161041f81610513565b50565b600061042c6105bc565b905090565b3660008037600080366000845af43d6000803e808015610450573d6000f35b3d6000fd5b606083156104c15782516104ba576001600160a01b0385163b6104ba5760405162461bcd60e51b815260206004820152601d60248201527f416464726573733a2063616c6c20746f206e6f6e2d636f6e7472616374000000604482015260640161015e565b50816104cb565b6104cb83836105e4565b949350505050565b6104dc8161060e565b6040516001600160a01b038216907fbc7cd75a20ee27fd9adebab32041f755214dbc6bffa90cc0225b39da2e5c2d3b90600090a250565b6001600160a01b0381166105785760405162461bcd60e51b815260206004820152602660248201527f455243313936373a206e65772061646d696e20697320746865207a65726f206160448201526564647265737360d01b606482015260840161015e565b807fb53127684a568b3173ae13b9f8a6016e243e63b6e8ee1178d6a717850b5d61035b80546001600160a01b0319166001600160a01b039290921691909117905550565b60007f360894a13ba1a3210667c828492db98dca3e2076cc3735a920a3ca505d382bbc6101db565b8151156105f45781518083602001fd5b8060405162461bcd60e51b815260040161015e9190610823565b6001600160a01b0381163b61067b5760405162461bcd60e51b815260206004820152602d60248201527f455243313936373a206e657720696d706c656d656e746174696f6e206973206e60448201526c1bdd08184818dbdb9d1c9858dd609a1b606482015260840161015e565b807f360894a13ba1a3210667c828492db98dca3e2076cc3735a920a3ca505d382bbc61059b565b600080858511156106b257600080fd5b838611156106bf57600080fd5b5050820193919092039150565b80356001600160a01b03811681146106e357600080fd5b919050565b6000602082840312156106fa57600080fd5b61019e826106cc565b634e487b7160e01b600052604160045260246000fd5b6000806040838503121561072c57600080fd5b610735836106cc565b9150602083013567ffffffffffffffff8082111561075257600080fd5b818501915085601f83011261076657600080fd5b81358181111561077857610778610703565b604051601f8201601f19908116603f011681019083821181831017156107a0576107a0610703565b816040528281528860208487010111156107b957600080fd5b8260208601602083013760006020848301015280955050505050509250929050565b60005b838110156107f65781810151838201526020016107de565b838111156103c45750506000910152565b600082516108198184602087016107db565b9190910192915050565b60208152600082518060208401526108428160408501602087016107db565b601f01601f1916919091016040019291505056fe416464726573733a206c6f772d6c6576656c2064656c65676174652063616c6c206661696c6564a264697066735822122012bb4f564f73959a03513dc74fc3c6e40e8386e6f02c16b78d6db00ce0aa16af64736f6c63430008090033b53127684a568b3173ae13b9f8a6016e243e63b6e8ee1178d6a717850b5d6103416464726573733a206c6f772d6c6576656c2064656c65676174652063616c6c206661696c65640000000000000000000000007e22fcb742572515d1c3fef972ec066c995820ef000000000000000000000000d89f6892a2b58d50e0e8f51a961b12a25ce3f7570000000000000000000000000000000000000000000000000000000000000060000000000000000000000000000000000000000000000000000000000000020457405d050000000000000000000000005f4ec3df9cbd43714fe2740f5e3616155c5b8419000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec70000000000000000000000000000000000000000000000000000000064c7eb930000000000000000000000000000000000000000000000000000000064d2bde300000000000000000000000000000000000000000000000000000000000000e0000000000000000000000000000000000000000000000000000000003b9aca000000000000000000000000004580900be782a567ad407cc8fcde24dbda2a1aae000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000a000000000000000000000000000000000000000000000000000000000000000e000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000165a0bc00000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000009795118749000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000023a8f4b6a000000000000000000000000000000000000000000000000000000000100ad53fa19000000000000000000000000000000000000000000000000000000000300979db18107552faa36a52480a1dbb65ed3f51d706fd00841d705c002a846de1b56d0c36941cd2897a7bb19419d427a252bc4fff45be5ed4725f8b6047b5309da5f47d1e98fa2c79e5c0841595291fae7167de10800"),
-        ssz_receipt_bytes=bytes.fromhex("3e0000007afb0e0000000000451445a9adb7e4c8ab4a266755626c136e093a4e0000000000000000000000000000000040000000000000000080000001000000000000000000000000000000000000000000000200004000000000000201000000000000000000000000000000000200000100000000000000000000000000000000000002000000000000000000080004000080000000000000000000000040000000000000000000000004000000000000000000008000000000040080000000000000000000000000000000040000000000000000000000000000000000000000002020000000000000000004000a00000000040000000000000000002000000000000000000000000000000000000000000000000000000000000002000021010000011400000070000000ec00000088010000e4010000451445a9adb7e4c8ab4a266755626c136e093a4e1c0000005c000000bc7cd75a20ee27fd9adebab32041f755214dbc6bffa90cc0225b39da2e5c2d3b0000000000000000000000007e22fcb742572515d1c3fef972ec066c995820ef451445a9adb7e4c8ab4a266755626c136e093a4e1c0000007c0000008be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000979db18107552faa36a52480a1dbb65ed3f51d70451445a9adb7e4c8ab4a266755626c136e093a4e1c0000003c00000023f6ad8232d75562dd1c6b37dfc895af6bfc1ecd0fb3b88722c6a5e6b4dc9a200000000000000000000000000000000000000000000000000000000064c7eb930000000000000000000000000000000000000000000000000000000064d2bde30000000000000000000000000000000000000000000000000000000064c7e277451445a9adb7e4c8ab4a266755626c136e093a4e1c0000003c0000007f26b83ff96e1f2b6a682f133852f6798a09c465da95921460cefb38474024980000000000000000000000000000000000000000000000000000000000000001451445a9adb7e4c8ab4a266755626c136e093a4e1c0000003c0000007e644d79422f17c01e4894b5f4f588d331ebfa28653d42ae832dc59e38c9798f0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000d89f6892a2b58d50e0e8f51a961b12a25ce3f757"),
     ),
 ]
 
-transactions = [upgrade_rlp_transaction_to_ssz(test.rlp_tx_bytes) for test in tests]
-receipts = upgrade_rlp_receipts_to_ssz([test.rlp_receipt_bytes for test in tests], transactions)
+for test in tests:
+    tx = upgrade_rlp_transaction_to_ssz(test.rlp_tx_bytes)
 
-for i in range(len(tests)):
-    validate_transaction(transactions[i])
-    assert compute_tx_hash(transactions[i]) == tests[i].rlp_tx_hash
-    assert transactions[i].hash_tree_root() == tests[i].ssz_tx_root
+    validate_tx_from_address(tx)
+    assert compute_tx_hash(tx) == test.rlp_tx_hash
+    assert tx.hash_tree_root() == test.ssz_tx_root
 
-    stable_transaction = Transaction(backing=transactions[i].get_backing())
-    assert stable_transaction.encode_bytes() == tests[i].ssz_tx_bytes
-    assert stable_transaction.hash_tree_root() == tests[i].ssz_tx_root
-
-    stable_receipt = Receipt(backing=receipts[i].get_backing())
-    assert stable_receipt.encode_bytes() == tests[i].ssz_receipt_bytes
-
-transactions_stable = [Transaction(backing=transaction.get_backing()) for transaction in transactions]
-assert select_transaction_profile(transactions_stable[0]) is ReplayableTransaction
-assert select_transaction_profile(transactions_stable[1]) is ReplayableTransaction
-assert select_transaction_profile(transactions_stable[2]) is LegacyTransaction
-assert select_transaction_profile(transactions_stable[3]) is Eip2930Transaction
-assert select_transaction_profile(transactions_stable[4]) is Eip1559Transaction
-assert select_transaction_profile(transactions_stable[5]) is Eip1559Transaction
-
-receipts_stable = [Receipt(backing=receipt.get_backing()) for receipt in receipts]
-assert select_receipt_profile(receipts_stable[0]) is HomesteadReceipt
-assert select_receipt_profile(receipts_stable[1]) is BasicReceipt
-assert select_receipt_profile(receipts_stable[2]) is BasicReceipt
-assert select_receipt_profile(receipts_stable[3]) is BasicReceipt
-assert select_receipt_profile(receipts_stable[4]) is BasicReceipt
-assert select_receipt_profile(receipts_stable[5]) is BasicReceipt
+    stable_tx = Transaction(backing=tx.get_backing())
+    assert identify_transaction_profile(stable_tx) is test.tx_profile
+    assert stable_tx.hash_tree_root() == test.ssz_tx_root
+    assert stable_tx.encode_bytes() == test.ssz_tx_bytes
